WIBI-400. AvroKeyValueInputFormat.
An input format that reads from Avro container files.  The Avro
container files are required to contain generic records, each with
exactly two fields: 'key' and 'value'.  The key and value fields
from the generic record will be read as the MR job's input key
and value as expected.

This reads files generated by the AvroKeyValueOutputFormat.